Fix downloading datasets Round vales of prettydate getScipionScript returns -m scipion

+++ pyworkflow/utils/utils.py
@@ -45,7 +45,7 @@ def prettyDate(time=False):
         diff = now - datetime.fromtimestamp(time)
     elif type(time) is float:
         diff = now - datetime.fromtimestamp(int(time))
-    elif isinstance(time,datetime):
+    elif isinstance(time, datetime):
         diff = now - time 
     elif not time:
         diff = now - now
@@ -61,22 +61,22 @@ def prettyDate(time=False):
         if second_diff < 60:
             return str(second_diff) + " seconds ago"
         if second_diff < 120:
-            return  "a minute ago"
+            return "a minute ago"
         if second_diff < 3600:
-            return str( second_diff / 60 ) + " minutes ago"
+            return str(int(second_diff/60)) + " minutes ago"
         if second_diff < 7200:
             return "an hour ago"
         if second_diff < 86400:
-            return str( second_diff / 3600 ) + " hours ago"
+            return str(int(second_diff/3600)) + " hours ago"
     if day_diff == 1:
         return "Yesterday"
     if day_diff < 7:
         return str(day_diff) + " days ago"
     if day_diff < 31:
-        return str(day_diff/7) + " weeks ago"
+        return str(int(day_diff/7)) + " weeks ago"
     if day_diff < 365:
-        return str(day_diff/30) + " months ago"
-    return str(day_diff/365) + " years ago"
+        return str(int(day_diff/30)) + " months ago"
+    return str(int(day_diff/365)) + " years ago"
 
 
 def dateStr(dt=None, time=True, secs=False, dateFormat=None):
